+++ CreditsSkipMonitor.cs
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: MIT
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;